+++ src/rogallo/data/config.py
@@ -61,6 +61,15 @@ class Configuration:
     cache_ttl: int = 3_600
     """The time-to-live for cached content, in seconds."""
 
+    connection_timeout: int = 10
+    """The connection timeout for network requests, in seconds."""
+
+    read_timeout: int = 30
+    """The read timeout for network requests, in seconds."""
+
+    maximum_redirects: int = 5
+    """The maximum number of redirects to follow for network requests."""
+
 
 ##############################################################################
 def configuration_file() -> Path: